chore: move `EnvVars/Nep330BuildInfo` to `crate::types::docker_build` modules
